pyproject.toml
[project]
name = "atlassian-cli"
version = "0.1.0"
description = "CLI wrapper around atlassian-python-api for Jira and Confluence"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "click>=7.0",
  "atlassian-python-api>=3.14.0"
]

[project.scripts]
atlassian-cli = "atlassian_cli.cli:cli"
